refactor(game): 사용자 입력을 InputView로 분리
