+++ src/main/java/racingcar/Application.java
@@ -1,7 +1,10 @@
 package racingcar;
 
+import racingcar.controller.Game;
+import racingcar.view.InputView;
+
 public class Application {
     public static void main(String[] args) {
-        new Game().play();
+        new Game(new InputView());
     }
 }
