+++ src/main/java/racingcar/controller/Game.java
@@ -1,26 +1,27 @@
-package racingcar;
+package racingcar.controller;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import racingcar.model.Car;
+import racingcar.model.Referee;
+import racingcar.view.InputView;
 
 public class Game {
-    private static final String INPUT_CAR_NAME_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
-    private static final String INPUT_COUNT_MESSAGE = "시도할 회수는 몇회인가요?";
+    private final InputView inputView;
+
+    public Game(InputView inputView) {
+        this.inputView = inputView;
+    }
+
     private static final String GAME_RESULT = "실행 결과";
     private static final String WINNER = "최종 우승자 : ";
 
-    private static final String SEPARATOR = ",";
 
     public void play() {
-        System.out.println(INPUT_CAR_NAME_MESSAGE);
-        String input = Console.readLine();
-        List<Car> cars = splitToCarList(input);
-
-        System.out.println(INPUT_COUNT_MESSAGE);
-        int count = Integer.parseInt(Console.readLine());
-        System.out.println();
+        List<Car> cars = inputView.inputCarName();
+        int count = inputView.inputCount();
 
         System.out.println(GAME_RESULT);
         for (int i = 0; i < count; i++) {
@@ -33,11 +34,7 @@ public class Game {
         System.out.println(extractWinnerName(winners));
     }
 
-    private List<Car> splitToCarList(String input) {
-        return Stream.of(input.split(SEPARATOR))
-                .map(Car::new)
-                .toList();
-    }
+
 
     private String extractWinnerName(List<Car> winners) {
         return winners.stream()
